requirements_streamlit.txt
streamlit
langchain-community
langchain-text-splitters
langchain-core
chromadb
sentence-transformers
beautifulsoup4
langchain-openai
